Commments are added and some warnings are removed

+++ src/main/java/brickGame/Bonus.java
@@ -16,12 +16,13 @@ public class Bonus implements Serializable {
     public boolean taken = false;
 
     public Bonus(int row, int column) {
-        x = (column * (Block.getWidth())) + Block.getPaddingH() + (Block.getWidth() / 2) - 15;
-        y = (row * (Block.getHeight())) + Block.getPaddingTop() + (Block.getHeight() / 2) - 15;
+        x = (column * (Block.getWidth())) + Block.getPaddingH() + ((double) Block.getWidth() / 2) - 15;
+        y = (row * (Block.getHeight())) + Block.getPaddingTop() + ((double) Block.getHeight() / 2) - 15;
 
         draw();
     }
 
+    //    Draws the bonus item, setting its appearance based on a random choice.
     private void draw() {
         choco = new Rectangle();
         choco.setWidth(30);
@@ -38,7 +39,6 @@ public class Bonus implements Serializable {
 
         choco.setFill(new ImagePattern(new Image(url)));
     }
-
 
 
 }
